// src/app/models/Product.ts
export interface Product {
    purchaseDate: string;
    invoice: string;
    customerRoot: string;
    customerLeaf: string;
    productDescription: string;
    packSize: string;
    untilType: string;
    category: string;
    distributorRoot: string;
    manafacturer: string;
    quantity: string;
    price: string;
    total: string;


}